% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compare.R
\name{compare_groups}
\alias{compare_groups}
\title{Compare groups between two data frames}
\usage{
compare_groups(x, y, group_cols)
}
\arguments{
\item{x, y}{Data frames to compare}

\item{group_cols}{<\code{\link[dplyr:dplyr_tidy_select]{tidy-select}}> Columns to use for grouping}
}
\value{
A data frame containing the grouping columns and two additional columns,
\code{in_x} and \code{in_y}, which are TRUE if the group values are in the corresponding
data frame and FALSE otherwise. Records where both \code{in_x} and \code{in_y} are TRUE
are excluded from the output.
}
\description{
Compare groups between two data frames
}
\examples{
x <- data.frame(team = c("a", "a", "b"), player = 1:3)
y <- data.frame(team = c("a", "b", "c"), player = 4:6)

# Team values that appear in only one of the frames
compare_groups(x, y, group_cols = team)
}
